Rename alpha3 multiclass files
