+++ neasqc_wp61/models/quantum/alpha/module/alpha_3_model.py
@@ -6,7 +6,7 @@ from pennylane import numpy as np
 
 # inherit from PennyLaneModel to use the PennyLane circuit evaluation
 class Alpha_3_model(PennyLaneModel):
-    def __init__(self, n_qubits, q_delta, device): 
+    def __init__(self, n_qubits, q_delta, n_classes, device): 
         """
         Definition of the *dressed* layout.
         """
@@ -16,11 +16,12 @@ class Alpha_3_model(PennyLaneModel):
         self.n_qubits = n_qubits
         self.q_delta = q_delta
         self.device = device
+        self.n_classes = n_classes
 
         self.pre_net = nn.Linear(768, self.n_qubits)
         self.q_params = nn.Parameter(self.q_delta * torch.randn((self.n_qubits + 2) * self.n_qubits))
-        self.post_net = nn.Linear(self.n_qubits, 2)
-        self.sigmoid = nn.Sigmoid()
+        self.post_net = nn.Linear(self.n_qubits, self.n_classes)
+        self.softmax = nn.Softmax()
 
         
         dev = qml.device('default.qubit', wires=self.n_qubits)
@@ -34,24 +35,24 @@ class Alpha_3_model(PennyLaneModel):
         net.
         """
 
-        # obtain the input features for the quantum circuit
-        # by reducing the feature dimension from 512 to 4
-        pre_out = self.pre_net(input_features)
-        q_in = torch.tanh(pre_out) * np.pi / 2.0
+       # obtain the input features for the quantum circuit
+        # by reducing the feature dimension from 768 to self.n_qubits
+        pre_out = self.pre_net(input_features)  #Here input_features.shape=(batch_size, 768) and pre_out.shape=(batch_size, self.n_qubits)
+        q_in = torch.tanh(pre_out) * np.pi / 2.0    #Here q_in.shape=(batch_size, self.n_qubits) and have values between -pi/2 and pi/2
         
         
-        q_out = self.quantum_net(q_in, self.q_params)
-        q_out = torch.stack(q_out)
-        q_out = q_out.transpose(0, 1).float()  # Transpose dimensions 0 and 1
+        q_out = self.quantum_net(q_in, self.q_params) #Output q_out a list with the shape (self.n_qubits, batch_size)
+        q_out = torch.stack(q_out)  # Stack outputs into one tensor of shape (self.n_qubits, batch_size)
+        q_out = q_out.transpose(0, 1).float()  # Transpose dimensions 0 and 1 to get a tensor of shape (batch_size, self.n_qubits)
 
         # return the two-dimensional prediction from the postprocessing layer
-        return self.sigmoid(self.post_net(q_out))
+        return self.softmax(self.post_net(q_out))
     
 
 
     def quantum_net_from_paper(self, inputs, q_weights):
         #https://onlinelibrary.wiley.com/doi/epdf/10.1002/qute.201900070
-        #Circuit [6] but adapted for n_qubits = 3
+        #Circuit [6] but adapted for n_qubits
 
 
         qml.AngleEmbedding(features=inputs, wires=range(self.n_qubits), rotation='X')
